+++ packages/dart_edge_sql/lib/src/core/sql_schema.dart
@@ -1,2 +1,2 @@
 export 'package:dart_edge_core/dart_edge_core.dart'
-    show SqlColumn, SqlRawTable, SqlTable;
+    show SqlColumn, SqlRawTable, SqlRowColumnRead, SqlTable;